+++ Helper/DateTimeHelper.php
@@ -33,13 +33,17 @@ class DateTimeHelper
      *
      * @param string $datetime Expected value in the Format 'YYYY-MM-DD 00:00:00'
      * @param string $outputFormat [optional] the Date will look like '2020-01-12T23:59:21+00:00' ('c' = ISO 8601)
+     * @param \DateTimeZone|null $timezone [optional] self::TIMEZONE_DEFAULT is used by default
      *
      * @return string
      */
-    public static function getFormattedValue(string $datetime, string $outputFormat = 'c'): string
+    public static function getFormattedValue(string $datetime, string $outputFormat = 'c', ?\DateTimeZone $timezone = null): string
     {
         try {
-            $date = new \DateTime($datetime, self::getTimezone());
+            if ($timezone === null) {
+                $timezone = self::getTimezone();
+            }
+            $date = (new \DateTime($datetime))->setTimezone($timezone);
         } catch (\Exception $e) {
             return $datetime;
         }
@@ -53,13 +57,17 @@ class DateTimeHelper
      * @param string $interval see https://www.php.net/manual/en/dateinterval.construct.php
      * @param string $dateBegin Start Date to add interval
      * @param string $format output date format
+     * @param \DateTimeZone|null $timezone [optional] self::TIMEZONE_DEFAULT is used by default
      *
      * @return string
      */
-    public static function addDateInterval(string $interval, string $dateBegin = 'now', string $format = 'c'): string
+    public static function addDateInterval(string $interval, string $dateBegin = 'now', string $format = 'c', ?\DateTimeZone $timezone = null): string
     {
         try {
-            $date = new \DateTime($dateBegin, self::getTimezone());
+            if ($timezone === null) {
+                $timezone = self::getTimezone();
+            }
+            $date = (new \DateTime($dateBegin))->setTimezone($timezone);
             $date->add(new \DateInterval($interval));
         } catch (\Exception $e) {
             // FIXME: WORKAROUND to proceed with correct intervals and return valid value
@@ -79,13 +87,17 @@ class DateTimeHelper
      * @param string $interval see https://www.php.net/manual/en/dateinterval.construct.php (Subtrahend)
      * @param string $dateBegin Start Date to subtract interval (Minuend)
      * @param string $format output date format
+     * @param \DateTimeZone|null $timezone [optional] self::TIMEZONE_DEFAULT is used by default
      *
      * @return string
      */
-    public static function subtractDateInterval(string $interval, string $dateBegin = 'now', string $format = 'c'): string
+    public static function subtractDateInterval(string $interval, string $dateBegin = 'now', string $format = 'c', ?\DateTimeZone $timezone = null): string
     {
         try {
-            $date = new \DateTime($dateBegin, self::getTimezone());
+            if ($timezone === null) {
+                $timezone = self::getTimezone();
+            }
+            $date = (new \DateTime($dateBegin))->setTimezone($timezone);
             $date->sub(new \DateInterval($interval));
         } catch (\Exception $e) {
             // FIXME: WORKAROUND to proceed with correct intervals and return valid value
@@ -104,13 +116,17 @@ class DateTimeHelper
      * NOTE: the result is based on self::getTimezone
      *
      * @param string $datetime Expected value in the Format 'YYYY-MM-DD 00:00:00'
+     * @param \DateTimeZone|null $timezone [optional] self::TIMEZONE_DEFAULT is used by default
      *
      * @return int {timestamp}
      */
-    public static function getTimestamp(string $datetime): int
+    public static function getTimestamp(string $datetime, ?\DateTimeZone $timezone = null): int
     {
         try {
-            return (new \DateTime($datetime, self::getTimezone()))->getTimestamp();
+            if ($timezone === null) {
+                $timezone = self::getTimezone();
+            }
+            return (new \DateTime($datetime))->setTimezone($timezone)->getTimestamp();
         } catch (\Exception $e) {
             return 0;
         }